Convert HQx and scrap to 32-bit color.
Allow putting small 32-bit pics onto the scrap and upscaling them.
Also limit upscaling level based on maximum GL rexture size.

+++ src/refresh/gl/gl.h
@@ -488,6 +488,6 @@ void GL_DrawAliasModel(model_t *model);
  * hq2x.c
  *
  */
-void HQ2x_Render(uint32_t *output, const uint8_t *input, int width, int height);
-void HQ4x_Render(uint32_t *output, const uint8_t *input, int width, int height);
+void HQ2x_Render(uint32_t *output, const uint32_t *input, int width, int height);
+void HQ4x_Render(uint32_t *output, const uint32_t *input, int width, int height);
 void HQ2x_Init(void);
